[Fix #47] room 페이지의 layout 구성 및 컴포넌트 분리
